requirements.txt
cycler==0.10.0
ete3==3.0.0b35
graphviz==0.5.1
matplotlib==1.5.3
numpy==1.11.2
Pillow==3.4.2
pydot==1.2.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
